\name{NewLazy-package}
\alias{NewLazy-package}
\alias{NewLazy}
\docType{package}
\title{Package of LazyCD}
\packageTitle{NewLazy}
}
\description{
\packageDescription{NewLazy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NewLazy}
\packageIndices{NewLazy}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NewLazy}

Maintainer: \packageMaintainer{NewLazy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
